requirements.txt
python-dotenv
pydantic
pypdf
openai
fastapi
uvicorn